+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.9.11](https://github.com/CalvinVon/dalao-proxy/compare/v0.9.0...v0.9.10) (2019-10-10)
+
+### BREAKING CHANGES
+- option `responseFilter`: default value changes from `["code", 200]` to `[]`.
+
+### Features
+- option `headers` support freely specify setting request header or response header ([#headers](./README.md#option-headers)).
+
 ## [0.9.10](https://github.com/CalvinVon/dalao-proxy/compare/v0.9.0...v0.9.10) (2019-07-26)
 
 ### Features